fastapi/requirements.txt
boto3==1.18.61
Pillow==8.3.1
numpy==1.21.0
fastapi==0.68.0
mangum==0.12.2
uvicorn==0.15.0
python-multipart==0.0.5